refactor: move pkg/ to internal/
