+++ db/sqlc.yaml
@@ -8,7 +8,7 @@ sql:
     gen:
       go:
         package: "gen"
-        out: "../pkg/infrastructure/oidc/gen"
+        out: "../internal/infrastructure/oidc/gen"
         emit_json_tags: true
         emit_prepared_queries: true
         emit_interface: true
@@ -21,7 +21,7 @@ sql:
     gen:
       go:
         package: "gen"
-        out: "../pkg/infrastructure/portal/gen"
+        out: "../internal/infrastructure/portal/gen"
         emit_json_tags: true
         emit_prepared_queries: true
         emit_interface: true
